Enable W4 warning level on MSVC. fix all (all-but-two) warnings.

+++ src/DGLNet/server.cpp
@@ -20,7 +20,7 @@
 
 namespace dglnet {
 
-    Server::Server(int port, MessageHandler* handler):Transport(handler),m_endpoint(boost::asio::ip::tcp::v4(), port), m_acceptor(m_io_service) {
+    Server::Server(unsigned short port, MessageHandler* handler):Transport(handler),m_endpoint(boost::asio::ip::tcp::v4(), port), m_acceptor(m_io_service) {
         m_acceptor.open(m_endpoint.protocol());
         m_acceptor.set_option(boost::asio::ip::tcp::acceptor::reuse_address(true));
         m_acceptor.bind(m_endpoint);
